pyproject.toml
[tool.poetry]
name = "fastapi-sqlalchemy-asyncpg"
version = "0.0.7"
description = ""
authors = ["Jakub Miazek <the@grillazz.com>"]
packages = []
license = "MIT"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "0.104.0"
pydantic = {version = "2.4.2", extras = ["email"]}
pydantic-settings = "2.0.3"
sqlalchemy = "2.0.22"
uvicorn = { version = "0.23.2", extras = ["standard"]}
asyncpg = "0.28.0"
httpx = "0.25.0"
pytest-cov = "4.1.0"
pytest-asyncio = "0.21.1"
uvloop = "0.18.0"
httptools = "0.6.1"
rich = "13.6.0"
pyjwt = {version = "2.8.0", extras = ["cryptography"]}
redis = "5.0.1"
passlib = {version = "^1.7.4", extras = ["bcrypt"]}
sourcery = "^1.11.0"
pandas = "2.1.4"

[tool.poetry.dev-dependencies]



[tool.poetry.group.dev.dependencies]
tryceratops = "^1.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
indent-width = 4

select = ["E", "F", "UP", "N", "C", "B"]
ignore = ["E501"]

# Exclude a variety of commonly ignored directories.
exclude = ["alembic",]
# Assume Python 3.10.
target-version = "py311"

[tool.ruff.flake8-quotes]
docstring-quotes = "double"

[tool.ruff.flake8-bugbear]
extend-immutable-calls = ["fastapi.Depends",]

[tool.pytest.ini_options]
addopts = "-v --doctest-modules --doctest-glob=*.md --ignore=alembic"
asyncio_mode = "strict"
env_files = [".env"]

[tool.tryceratops]
exclude = ["alembic",]

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"